refactor: Remove deprecated socket and TCA feature files
Remove obsolete SocketClient.swift, TCAFeatureAction.swift, and WebRTCSocket components that have been replaced by the WebRTCCore package architecture.
